Avoid constructing invalid Contents in SendMessage.
`ChatSession#SendMessage` currently unconditionally adds an entry to the ChatSession.
This is illegal if no parts were actually provided - the resulting object will be
rejected by the API server. Additionally, always assuming that the content type is
"user" is limiting.

Since the current behaviour is always illegal, don't bother constructing an object
if the caller didn't provide any parts to append.